Adds a list to help with game function
Uses list to represent what box the cookie is in, so the broccoli does not have the same box.

Can also set up a variable that has value of cookie index and uses a conditional to check if its the same index as cookie.
This has the possiblity of running multiple times if the rnd number generated gets the same index more than once.
Do a speed test to see which way is faster and what the benefits and drawback are to each